Extract parsing helpers into shared module
