+++ src/parser/mod.rs
@@ -1221,7 +1221,7 @@ pub mod ast {
         None
     }
 
-    fn skip_whitespace_and_comments<I>(iter: &mut std::iter::Peekable<I>)
+    pub(super) fn skip_whitespace_and_comments<I>(iter: &mut std::iter::Peekable<I>)
     where
         I: Iterator<Item = SyntaxElement<DdlogLanguage>>,
     {
@@ -1233,105 +1233,8 @@ pub mod ast {
         }
     }
 
-    fn parse_name_type_pairs<I>(mut iter: I) -> Vec<(String, String)>
-    where
-        I: Iterator<Item = SyntaxElement<DdlogLanguage>>,
-    {
-        use rowan::NodeOrToken;
-
-        // Skip to the first '('
-        let mut depth = 0usize;
-        for e in &mut iter {
-            if e.kind() == SyntaxKind::T_LPAREN {
-                break;
-            }
-        }
-
-        let mut pairs = Vec::new();
-        let mut buf = String::new();
-        let mut name: Option<String> = None;
-        for e in iter {
-            match e {
-                NodeOrToken::Token(t) => match t.kind() {
-                    SyntaxKind::T_LPAREN => {
-                        depth += 1;
-                        buf.push_str(t.text());
-                    }
-                    SyntaxKind::T_RPAREN => {
-                        if depth == 0 {
-                            if let Some(n) = name.take() {
-                                let ty = buf.trim();
-                                if !ty.is_empty() {
-                                    pairs.push((n, ty.to_string()));
-                                }
-                            }
-                            break;
-                        }
-                        depth -= 1;
-                        buf.push_str(t.text());
-                    }
-                    SyntaxKind::T_COMMA if depth == 0 => {
-                        if let Some(n) = name.take() {
-                            let ty = buf.trim();
-                            if !ty.is_empty() {
-                                pairs.push((n, ty.to_string()));
-                            }
-                        }
-                        buf.clear();
-                    }
-                    SyntaxKind::T_COLON if depth == 0 => {
-                        name = Some(buf.trim().to_string());
-                        buf.clear();
-                    }
-                    _ => buf.push_str(t.text()),
-                },
-                NodeOrToken::Node(n) => buf.push_str(&n.text().to_string()),
-            }
-        }
-
-        pairs
-    }
-
-    fn parse_type_after_colon<I>(iter: &mut std::iter::Peekable<I>) -> Option<String>
-    where
-        I: Iterator<Item = SyntaxElement<DdlogLanguage>>,
-    {
-        use rowan::NodeOrToken;
-
-        skip_whitespace_and_comments(iter);
-        if !matches!(
-            iter.peek().map(SyntaxElement::kind),
-            Some(SyntaxKind::T_COLON)
-        ) {
-            return None;
-        }
-        iter.next();
-
-        let mut buf = String::new();
-        for e in iter {
-            match e {
-                NodeOrToken::Token(t)
-                    if matches!(t.kind(), SyntaxKind::T_LBRACE | SyntaxKind::T_SEMI) =>
-                {
-                    break;
-                }
-                NodeOrToken::Token(t) => {
-                    if t.kind() == SyntaxKind::T_WHITESPACE && t.text().contains('\n') {
-                        break;
-                    }
-                    buf.push_str(t.text());
-                }
-                NodeOrToken::Node(n) => buf.push_str(&n.text().to_string()),
-            }
-        }
-
-        let text = buf.trim();
-        if text.is_empty() {
-            None
-        } else {
-            Some(text.to_string())
-        }
-    }
+    mod parse_utils;
+    use parse_utils::{parse_name_type_pairs, parse_type_after_colon};
 
     /// Typed wrapper for a relation declaration.
     #[derive(Debug, Clone)]
